+++ src/components/interactive/ButtonGroup.jsx
@@ -5,20 +5,41 @@ import combine from '../../util/combine';
 
 const ButtonGroup = ({ buttons, onChange, ...restBtnGroup }) => (
   <div {...restBtnGroup} className="btngroup">
-    {buttons.map(({ value, content, key, Icon, onClick, label, ...restBtn }) => (
-      <GroupButton
-        key={key || label}
-        onClick={onChange ? combine(() => onChange(label || value), onClick) : onClick}
-        label={label}
-        Icon={Icon}
-        {...restBtn}
-      />
-    ))}
+    {buttons.map(
+      (
+        {
+          value,
+          content,
+          key,
+          Icon,
+          onClick,
+          label,
+          // `isFirstButton` and `isLastButton` are private and can't being passed in.
+          isFirstButton,
+          isLastButton,
+          ...restBtn
+        },
+        index
+      ) => (
+        <GroupButton
+          key={key || label}
+          onClick={onChange ? combine(() => onChange(label || value), onClick) : onClick}
+          label={label}
+          Icon={Icon}
+          isFirstButton={index === 0}
+          isLastButton={index === buttons.length - 1}
+          {...restBtn}
+        />
+      )
+    )}
   </div>
 );
 
 ButtonGroup.propTypes = {
-  /** Each object in array accepts value, content, onClick, and icon, and also will pass through any extra props */
+  /**
+   * Each object in array renders a GroupButton, and (generally) passes all it's values through
+   * as props to GroupButton
+   */
   buttons: PropTypes.arrayOf(
     PropTypes.shape({
       /**
@@ -31,11 +52,26 @@ ButtonGroup.propTypes = {
        * Use if labels are not unique
        */
       key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+
       isActive: PropTypes.bool,
+      /**
+       * The text that appears in the button
+       */
       label: PropTypes.string,
+      /** Used to render a FDS Icon (should only be used for FDS Icons) */
       Icon: PropTypes.func,
+      /**
+       * A component to wrap the entire group button. The
+       * `wrapper` must render the children prop passed into it
+       * to render the GroupButton.
+       */
+      Wrapper: PropTypes.elementType,
     })
   ).isRequired,
+
+  /**
+   * Fires when any Button is clicked in ButtonGroup
+   */
   onChange: PropTypes.func,
 };